backend-data-processor/requirements.txt
# API Framework
fastapi==0.115.5
uvicorn[standard]==0.34.0
pydantic==2.10.3
pydantic-settings==2.6.1

# Async support
aiofiles==23.2.1
httpx==0.26.0

# Message Queue (Local simulation)
redis==5.0.1

# Database (Local Firestore emulator)
google-cloud-firestore==2.14.0

# Database (Local SQLite - alternative)
aiosqlite==0.19.0

# Data processing
python-multipart==0.0.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Code quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Monitoring & Logging
python-json-logger==2.0.7

# Development
python-dotenv==1.0.0